First attempt to export parametric models
Note that this is non trivial in general, since parameters in the
input model may correspond to complex expressions in the exported
model, due to differences e.g. in the representation of orientation.

A possibility is simply to generate a parameter whenver an expression
of the input parameters would be needed (maybe with a comment showing
the complete expression).