Make pygraphviz an optional dependency
Because it is often a pain in the neck to install due to the
C library (especially in Windows, cygwin, etc.), and it is
only needed to generate the .dot file.

The help string of the 'rmt dot' command reminds of the
dependency, and 'rmt dot' fails with an exception when
pygraphviz is not there.

Added as an optional dependency in the project metadata.