Log an error when a parameter value is not available
Instead of crashing due to a dictionary lookup miss.

This is in the routine that resolves the parameters in
a model.

A missing value will probably cause some other error
downstream anyway, so I might have used an exception
instead of just an error log. But I preferred to be
optimistic.