Add support for joint limits information
The idea is to be able to compose joint-limits data
with an input model format that does not account for
it (e.g. KinDSL), to export to a model format that
does (e.g. URDF).

Added new dataclass and model class to represent
joint limits, and command line option to specify
an extra input file with joint limits.

URDF export was modified to take this new model
into account, when available.

I also added a new Python module to group the
functions to load input models, to move them
out of the rmt module which is a bit too messy.
So far only loading of joint limits, though.